Refactor Symfony Serialization
To Support complex message graph the EnvelopeNormalizer must ask the
serializer to denormaliza as there can be other denormalizers. To
accomplish this a DefaultMessageNormalizer was created.

Also to have a more complete example of custom deserialization a Fixture
directory with a SendNewsletterMessage is added with custom metadata for
JMS Serializer (Annotations) and Symfony Serializer Component (CustomNormalizer).