MongoDB storage driver with unit/functional tests
This commit builds upon @iampersistent's work in bernardphp/bernard#94.

The driver is now constructed with MongoCollection objects, which removes the previous MongoDB::selectCollection() logic. Additionally, the message sentAt field is now a BSON date and queue name was added to acknowledgeMessage()'s remove() criteria to be consistent with the SQL driver.