Fix round robin queue tests under PHP 7
PR #193 modified the Travis configuration to disallow test failures
under PHP 7. This revealed some test failures for RoundRobinQueue,
which was recently added via PR #188, that are seemingly specific
to PHP 7.

These test failures don't appear to be caused by a defect of the
code under test, but seem likely to be due to a known regression
in PHPUnit's mocking library, namely sebastianbergmann/phpunit-mock-objects#267.

This change is a work-around for this regression. It modifies the
affected tests such that, instead of using full mock instances of
Queue for the queues contained by RoundRobinQueue, they instead use
partial mock instances of InMemoryQueue so as to use its
implementation of __toString() and thus avoid the need to stub it.
It's admittedly not ideal, but barring any significant changes to
InMemoryQueue, it should allow the RoundRobinQueue test suite to
pass and still remain accurate in the results it reports.